requirements.txt
asgiref==3.3.1
astroid==2.2.5
autopep8==1.4.4
certifi==2019.11.28
cffi==1.14.0
colorama==0.4.1
Django==3.1.4
django-crispy-forms==1.10.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-suit @ https://github.com/darklow/django-suit/tarball/v2
django-taggit==1.2.0
isort==4.3.20
lazy-object-proxy==1.4.1
Markdown==3.1.1
markdown2==2.3.1
mccabe==0.6.1
pi==0.1.2
Pillow==7.1.2
Pillow-PIL==0.1.dev0
pipenv==2018.11.26
psycopg2==2.8.4
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pycparser==2.20
pylint==2.3.1
pytz==2020.4
six==1.12.0
sqlparse==0.4.1
typed-ast==1.4.0
virtualenv==16.4.3
virtualenv-clone==0.5.3
wrapt==1.11.1
